requirements.txt

# ter baixado no vs code

google-generativeai>=0.3.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pillow>=10.0.0
python-dotenv>=1.0.0